LayoutTests/imported/w3c/web-platform-tests/css/css-variables/var-ident-function-expected.txt


FAIL Referencing a custom property with ident() assert_equals: expected "PASS" but got "FAIL1"
PASS ident() remains unresolved on custom properties
PASS ident() causing lookup of invalid custom property
FAIL ident() causing lookup of invalid custom property, fallback assert_equals: expected "PASS" but got ""
FAIL ident() causing lookup of invalid custom property, fallback, CSS-wide keyword assert_equals: expected "PASS" but got "FAIL"
